[refactoring] 리뷰 작성기능 로직 외부로 분리  #108
